Change due to refactoring of cypher internals
A lot of classes has been moved and hence the docs-build needs to be updated.
