+++ cypher/cypher-docs/src/test/scala/org/neo4j/cypher/docgen/QueryPlanTest.scala
@@ -22,7 +22,7 @@ package org.neo4j.cypher.docgen
 import org.hamcrest.CoreMatchers._
 import org.junit.Assert._
 import org.junit.Test
-import org.neo4j.cypher.internal.compiler.v3_3.pipes.IndexSeekByRange
+import org.neo4j.cypher.internal.compatibility.v3_3.runtime.pipes.IndexSeekByRange
 import org.scalatest.Ignore
 
 class QueryPlanTest extends DocumentingTestBase with SoftReset {
