+++ cypher/cypher-docs/src/test/scala/org/neo4j/cypher/docgen/tooling/DocumentingTest.scala
@@ -23,7 +23,7 @@ import java.io._
 import java.net.InetSocketAddress
 import java.util.Collections
 
-import org.neo4j.cypher.internal.compiler.v3_3.CypherSerializer
+import org.neo4j.cypher.internal.compatibility.v3_3.runtime.CypherSerializer
 import org.neo4j.cypher.internal.frontend.v3_3.test_helpers.CypherFunSuite
 import org.neo4j.cypher.internal.helpers.GraphIcing
 import org.neo4j.cypher.internal.spi.v3_3.TransactionBoundQueryContext.IndexSearchMonitor
