+++ cypher/cypher-docs/src/test/scala/org/neo4j/cypher/docgen/tooling/QueryResultContentBuilder.scala
@@ -19,8 +19,7 @@
  */
 package org.neo4j.cypher.docgen.tooling
 
-import org.neo4j.cypher.internal.compiler.v3_3.executionplan.InternalExecutionResult
-import org.neo4j.cypher.internal.frontend.v3_3.Rewritable._
+import org.neo4j.cypher.internal.compatibility.v3_3.runtime.executionplan.InternalExecutionResult
 import org.neo4j.cypher.internal.helpers.GraphIcing
 
 /**
